+++ dlls/player.h
@@ -234,7 +234,7 @@ public:
 	int					m_iDeaths;
 	float				m_flRespawnTimer;	// used in PlayerDeathThink() to make sure players can always respawn
 
-	int m_lastx, m_lasty;  // These are the previous update's crosshair angles, DON"T SAVE/RESTORE
+	float m_lastx, m_lasty;  // These are the previous update's crosshair angles, DON"T SAVE/RESTORE
 
 	int m_nCustomSprayFrames;// Custom clan logo frames for this player
 	float	m_flNextDecalTime;// next time this player can spray a decal
